backend/requirements.txt
astropy==6.0.0
astroquery==0.4.6
cryptography==41.0.5
gunicorn==21.2.0
dj-rest-auth==5.0.1
Django==4.2.6
django-cors-headers==4.3.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.2
pandas==2.1.4
psycopg==3.1.13
psycopg-binary==3.1.13
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.0
sqlparse==0.4.4
urllib3==2.1.0
